requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
python-dotenv   
Flask-Migrate
email-validator
pyjwt